refactor: components/board → components/menu 폴더 이름 변경
