+++ src/components/menu/MenuBoard/_BoardEmpty/BoardEmpty.tsx
