+++ src/components/menu/MenuBoard/_BoardEmpty/BoardEmpty.types.ts
