+++ src/components/menu/MenuBoard/_BoardEmpty/constants.ts
