+++ src/components/menu/MenuBoard/_CourseRow/CourseRow.styles.ts
