+++ src/components/menu/MenuBoard/_CourseRow/CourseRow.tsx
