+++ src/components/menu/MenuBoard/_CourseRow/CourseRow.types.ts
